requirements.txt
avro-python3
fastavro
requests-async==0.4.1
confluent-kafka==1.0.0

# Code quality
# ------------------------------------------------------------------------------
black==19.3b0
flake8==3.7.6

# Testing
# ------------------------------------------------------------------------------
tox==3.9.0
mypy==0.701
pytest==4.4.1
pytest_cases==1.6.2
pytest-asyncio==0.10.0